signed int omp fix

+++ trrojancl/include/trrojan/opencl/volume_raycast_benchmark.h
@@ -317,7 +317,7 @@ namespace opencl
             {
                 double div = pow(2.0, (sizeof(From) - sizeof(To))*8);
 #pragma omp parallel for
-                for (size_t i = 0; i < converted_data.size(); ++i)
+                for (long long int i = 0; i < converted_data.size(); ++i)
                 {
                     converted_data.at(i) = s[i] / div;
                 }
